+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Elementor widget changed from CTA to image box
+
 ### 1.0.4: September 23rd, 2020
 * Paragraph styling on mobile
 * Image width